+++ src/app/np-badge-demo/np-badge-demo-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpBadgeDemoComponent } from './np-badge-demo.component';
+
+const routes: Routes = [{
+  path: '',
+  component: NpBadgeDemoComponent
+}];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpBadgeDemoRoutingModule { }