+++ src/app/np-badge-demo/np-badge-demo.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { NpBadgeDemoRoutingModule } from './np-badge-demo-routing.module';
+import { NpBadgeDemoComponent } from './np-badge-demo.component';
+import { NpCardModule } from 'np-ui-lib';
+
+@NgModule({
+  declarations: [NpBadgeDemoComponent],
+  imports: [
+    CommonModule,
+    NpBadgeDemoRoutingModule,
+    NpCardModule
+  ]
+})
+export class NpBadgeDemoModule { }