+++ src/app/np-button-demo/np-button-demo-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpButtonDemoComponent } from './np-button-demo.component';
+
+const routes: Routes = [{
+  path: '',
+  component: NpButtonDemoComponent
+}];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpButtonDemoRoutingModule { }