+++ src/app/np-form-demo/np-form-demo-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpFormDemoComponent } from './np-form-demo.component';
+
+const routes: Routes = [{
+  path: '',
+  component: NpFormDemoComponent
+}];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpFormDemoRoutingModule { }